thes-2455: s3 dry-run support, updated cli, updated codefresh.yml

+++ src/aws/aws_s3.rs
@@ -1,23 +1,26 @@
+use crate::qrp::QrpFormat;
+use crate::utils::logging::get_root_logger;
+use anyhow::anyhow;
 use aws_sdk_s3::config::BehaviorVersion;
-use aws_sdk_s3::operation::put_object::PutObjectOutput;
 use aws_sdk_s3::primitives::ByteStream;
 use chrono::Utc;
 use clap::Parser;
-
-use crate::qrp::QrpFormat;
+use slog::info;
+use std::error::Error;
 
 #[derive(Parser, Debug, Clone)]
 pub struct AwsConf {
     #[clap(long, env)]
-    pub aws_endpoint: Option<String>,
-    #[clap(long, env)]
     pub qrp_bucket_name: String,
+    #[clap(long, env, default_value = "false")]
+    pub s3_dry_run: bool,
 }
 
 #[derive(Clone)]
 pub struct S3Client {
     aws_conf: AwsConf,
     client: aws_sdk_s3::Client,
+    log: slog::Logger,
 }
 
 impl S3Client {
@@ -26,6 +29,7 @@ impl S3Client {
         Ok(S3Client {
             aws_conf,
             client: aws_sdk_s3::Client::new(&config),
+            log: get_root_logger(),
         })
     }
 
@@ -35,19 +39,28 @@ impl S3Client {
         vat_number: &String,
         user: &String,
         format: QrpFormat,
-    ) -> anyhow::Result<PutObjectOutput> {
+    ) -> anyhow::Result<()> {
         let now = Utc::now();
         let date_time = now.format("%d_%m_%Y_%H:%M:%S");
         let lower_case_format = format.value();
         let file = format!("qrp/{vat_number}/{date_time}_{user}.{lower_case_format}");
 
+        if self.aws_conf.s3_dry_run {
+            info!(self.log, "Dry run: not uploading to S3");
+            return Ok(());
+        }
+
         Ok(self
             .client
             .put_object()
             .bucket(&self.aws_conf.qrp_bucket_name)
-            .key(file)
+            .key(&file)
             .set_body(Option::from(ByteStream::from(data.to_owned())))
             .send()
             .await?)
+        .and_then(|_res| {
+            info!(self.log, "Uploaded to S3"; "file" => &file);
+            Ok(())
+        })
     }
 }
